Create bank account
Small first step to help drive the initial core of the domain project.

Extract micro types for the account id and the amount. The latter is
not currency specific but just minor units, which is pence for British
pounds.

Externalise account id generation to avoid having static references
in the class, which would have made the test more difficult to write.