List bank accounts persisted in repository
Add the first port, a driven one, which allows to persist accounts.

The domain should only provide the contract definition of how it
interacts with the repository and the decision on how the mechanism is
going to work in the production system is delayed without interrupting
development. However, an in-memory implementation is available for test
purposes so that mocking is not necessary.