Implement bank http api
Use http4k library to implement http wrapper to the bank application.

This represent an adapter to the Bank interface, an implementation of
the driving port that defines the bank capabilities, and it trigger
actions on the application, the BankService.

Alongside the actual http application, create a client that can be used
externally to speak to the api, so that urls don't have to leak. Even
better, the client can implement the Bank contract so that tests are
implemented based on it.

Include 2 levels of testing, a simple in-memory test and an integration
test that spins up an http server.

As there's no intention to use a db or any other type of permanent
storage for this simple example, move the in-memory implementation of
the repository to production code so that it can be re-used.