Implement web ui using server templating
The only intention of this UI is to demonstrate how the Bank contract
can be used to drive its testing, expanding its coverage to the highest
level of abstraction.

As it's not meant to be production ready but just to serve as example,
the tests run in memory rather than against actual running services. It
has the drawback of coupling the web and http projects more than it
needs to.

On the implementation, it's important that the hbs template is named
after the ViewModel data class and is located in the same package
structure for http4k to recognise it.