Implement end to end account deposits
Start writing a new functional test to prove how a deposit should alter
an account's balance which drives the design of the system by updating
the Bank interface and its collaborators to deliver the solution.

The BankAccount domain model starts taking shape and needing to affect
the account state means that we should write a test for it. It also
affects the Amount which also has tests of its own.

Once the domain is updated, the subsequent layers need to be updated
to add the mappings to bring deposits to life.